fits: update Unmapper implementation for java 9+
The fits.Unmapper class uses classes outside of the public J2SE API
(using reflection) to perform explicit unmapping of mapped byte
buffers, used extensively for FITS handling.  This has worked for me
at Oracle java6, java7 and java8, but at java 9 the class it was using
(sun.misc.Cleaner) has disappeared from the Oracle JRE.

Following some investigation, it seems that the sun.misc.Unsafe
class can be used to do this at java9 and probably java10;
not sure about later versions.  So update the class to use
this strategy where required.  Modify the configuration and
automatic defaulting code as well since it was getting a bit
more complicated, but that essentially does the same as before:
lets you choose an option using a system property, or default
to something sensible.