auth: minimal customisation of build file
Replace placeholders in the build.xml file.
Comment various bits of functionality out since they are
not likely to be used, at least initially, for this package.
I followed the usage in package DPAC when doing this, since that
was a recently-added superficially similar library package.

By design this package has no dependencies, to facilitate its use
in third-party contexts.