ttools: note recent TAP-related changes in user docs
Summarise recent user-visible changes to TAP functionality resulting
from TAP-1.1 work: mostly taplint enhancements, but some changes
in TAP client requests too.  Although a lot of code has changed,
this is mostly internal reorganisation to allow multi-version TAP
support and to prepare for authentication implementation, so the
changes to TOPCAT and STILTS are not many.