ttools: fix log axis plot ranging bug
When data ranges are assessed for axis auto-ranging prior to a plot,
a small amount of padding is added at either end so that points near
the boundaries are not omitted from the plot.  However this wasn't
being done correctly in the case that the axis was logarithmic and
some of the values for the relevant coordinate were non-positive.
The result was that in some cases the auto-ranged plot could (just)
exclude points on the range boundary.  Fix it.

This bug was reported by Bertrand Goldman.