ttools: upgrade JEL to rev 179
This JEL version includes a fix for a JEL bug that resulted in unwanted
DEBUG output to stderr when a String function returned null.

This version corresponds to
http://svn.savannah.gnu.org/viewvc/jel?view=revision&revision=179
and will form part of JEL 2.1.2 when that is released.
I rebuilt it from source (ant dist) as usual, to get a java6-compatible
jel_g.jar file.