vo: change standard capability/interface constraints
Change constraint intf_type='vs:paramhttp' to intf_role='std'
when searching for standard services in the registry.
This is on advice from Markus Demleitner (see message on
{apps,registry}@ivoa.net 5 Sep 2019), and following RegTAP 1.1
and associated cleanup of registries.  There was a false start
a few months ago (this change would have missed some entries),
but I have checked searchable registry contents and they appear
now to comply with this new procedure.