topcat: add WebMapper to BrowserActivationType
The View In Web Browser activation type now lets you supply
things other than URLs or file names in the Location field;
a WebMapper is used to make sense of things like Bibcodes
and DOIs as well.  By default an "auto" mapper is used that
tries to makes sense of something that might be any of those
things.