ttools: attempt to handle @see tags in custom doclets
The doclets used to document the JEL functions from
uk.ac.starlink.ttools.funcs only handle a limited number
of doclet tags.  Add partial handling of the @see tag:
if that contains an HTML <a> element, it is now propgated
to the output in a suitable way for both XML and HTML backends.
@see tags pointing to other items on the classpath though
are still not handled.