votable: avoid deprecated use of auxiliary classes
A couple of the source files that were written a long time ago
have multiple package-visible classes at the top level,
used in other source files.  This practice provokes warnings at
java 8, so reorganise the files to avoid it.

These refactoring changes are not visible outside of the VOTABLE package.