fits: fixes to avoid compiler warnings
Sundry changes to avoid linter compiler warnings.
In some caes (ColumnReader) the class should really be changed
to be parameterised itself, but for now I'm not touching it,
and just wholesale s/Class/Class<?>/g to keep the compiler happy.

In a couple of places I have suppressed "cast" warnings, since
I consider that the redundant casts that were producing the warnings
actually serve to increase readability.