util: avoid build-time warnings
Sundry changes to the source code and javadocs to reduce the number
of compiler and javadoc warnings at java 8.

This includes slapping a SuppressWarnings annontations on a couple
of classes that use Swing classes that have acquired a parameterised
type between java 6 and java 8, so that the classes can still build
under either version but don't issue warnings at java 8.