table: use generics throughout
Java 8 now gives no compile-time warnings.
Most of the changes are adding parameterisations to parameterised types,
in most cases the code is not changed otherwise in a way that will not
be erased at compile time, but in a few cases some minor changes
have been made to accommodate the generics.  Various other warnings
like redundant casts have been avoided too.
A lot of lines of code are changed here.

SuppressWarnings is used in a few places, notably when dealing with
classes like JComboBox that have acquired generics since java 6.