vo: update VOLLT adqlParser.jar to 1.5-1
There was a bug in the tryQuickFix function of the VOLLT ADQL parser
library at v1.5 that often caused stack traces to stderr when there
was trailing whitespace in the ADQL expression being edited.
This commit replaces the library jar with v1.5-1, a bugfix release
issued by Gregory Mantelet.

As usual, this has been built from source by me with very minor changes
to get a Java6-compatible version.