hds: fix HDS/NDF unit test race condition
One unit test of the HdxFactory that usually worked (for me) at java 6
often fails (for me) at java 8.  This seems to come down to different
sequence of running the tests; a static initialiser was not being
called early enough to initialise the HdxFactory.  Put a suitable
line in the TestCase.setUp method to make sure this doesn't happen.