starjava: don't sign jars by default
Modify all the package build files so that the <signjar> tasks are
not by default executed as part of the install target.
The jars will be signed only if the "sign.jars" property is set
(e.g. ant -Dsign.jars=true).

Wholesale jar signing is a minority sport, only required when preparing
a WebStart installation.  In addition, the signing certificate used
by the default build-file controlled configuration is self-signed
and very out of date, hence of dubious value.  So the default is
now not to sign.

The SPLAT build file already had some logic along these lines,
but with the opposite sense; this commit replaces that by the same
mechanism now used in the other build files, so splat's
"skip.sign.jar" property is no longer used.