starjava: add more compiler checking to some build files
For files with source code that compiles sufficiently cleanly under
JDK8, add the javac flag

   -Xlint:all,-path,-serial

This warns about lots of things, usually what generates most warnings
is unchecked/rawtypes warnings as a consequence of incorrect or
missing use of generics.  The -path and -serial warnings opted out
of here don't seem to be very useful in the starjava context.

For packages whose use of generics (etc) has not so far been tidied up,
this lint flag has not yet been added, since the amount of noise it
generates is considerable.