jniast: fix HTML entities in javadoc
Fix the (mostly auto-generated) javadocs for JNIAST so that characters
'<', '>' and '&' are referred to as HTML character entities rather than
being inserted literally.  There are quite a lot of these, and this
error has been in JNIAST for a very long time, in some cases
(e.g. astResample) making the javadocs somewhat nonsensical in parts.

The changes here are both to the perl scripts, and as a result of
running those script using the build-source targets to regenerate
the auto-generated Java source files.