jniast: more javadoc appeasement
JNIAST javadocs have several miscellaneous headings that were
marked in the HTML with <H4> tags.  Java8 javadoc complains
about this and says they should be <H3> instead, so make the
relevant changes to avoid build noise.  The resulting HTML javadocs
tend to make these headings rather more prominent than they
probably ought to be, so some CSS changes would probably be
in order, but it's comprehensible as it is.