ttools: add taplint tests for TAP 1.1 capabilities
Add more taplint tests in stage CAP: test presence and arrangement
of capability, interface and securityMethod elements in the
capabilities document, in accordance with the requirements and
examples laid out in section 2.4 of PR-TAP-1.1-20190626.
Previously this stage only really tested the TAPRegExt requirements
of the capabilities document, but with TAP 1.1 the rest of it
is more important (security method declarations etc).