topcat: factor HTML rendering out of HtmlWindow class
Modify the implementation of the HtmlWindow class (a very simple
HTML browser used e.g. for the Browser activation type) so that
the actual HTML rendering is abstracted away to a new class
AbstractHtmlPanel.  A few details of behaviour are changed at
this commit, but the main point is to allow alternative HTML
rendering implementations.