topcat: HtmlWindow now has a JavaFX-based option
A new AbstractHtmlPanel concrete subclass FxHtmlPanel is introduced,
along with machinery for using it in HtmlWindow.
JavaFX is not guaranteed to be present in the JRE - I'm not sure
exactly what the rules are, but it's in my Oracle JDK8,
but I don't think it's typically in OpenJDK JREs.
So we have to be careful in construction to avoid nasty load errors.