topcat: provide JavaFX web activation browser
New option in the Browsers configurator for the BrowserActivationType
that uses the JavaFX-based internal HTML display window, if the
relevant runtime classes are available.

The JavaFX-based implementation, unlike the Swing-based one,
is capable of rendering javascript (and probably copes better with
a load of other things you might find on a modern web page as well).