fits: read offset long integers as Strings
Modify the way that 64-bit integer columns with a non-zero integer
offset are treated.  If TFORMn='K', TSCALn=1.0, and TZEROn is a
non-zero integer, then the range of possible values in the column
cannot be represented by a java long (the most common case
is using TFORMn='K' with TZEROn=9223372036854775808=2^63 to
represent unsigned 64-bit integers).  Up till now, this was dealt
with on a value-by-value basis, so that in-range values were treated
correctly, and out-of-range values simply looked like a null
(with a single WARNING issued through the logging system the
first time this happened).  But if you actually need e.g.
unsigned 64-bit values, this is not really good enough.

So implement a different option, where such columns are read
as String values.  They will not currently be round-tripped
correctly (the output handler will write Strings as well),
though that could be fixed.  Apart from FITS, I don't think any
supported table formats can accommodate offset/unsigned long values
(not VOTable or CDF, anyway), so there doesn't seem much point
propagating this outside of the FITS package.

This option (use Strings rather than best-efforts Longs) is
currently controlled by a static final flag in ColumnReader.
In principle it could be a configurable option of the FITS
input handler, but at present there's no API to change it.

This change was stimulated by a user reporting problems with
sometimes-blank values for the column SpecObjID from table in
SDSS DR15 tables from casjobs.  I don't really understand where
the FITS table in question came from though, because that column
is documented since DR14 as a 13-byte integer value, and when I
try to acquire data in FITS or VOTable I get string values anyway.