fits: support unsigned/offset long column output
If a String-valued table column is marked with the special metadata
value BintableStarTable.LONGOFF_INFO, then it will get output as
an offset long column (TFORMn='K', TZEROn=...).  This is the
opposite of what happens if such a column is encountered on read,
so it allows round-tripping of offset (e.g. unsigned) long values
in input FITS tables.