topcat: add Vizier beta servers to mirror list
The list of Vizier services that the Vizier load dialogue can query
now contains two new entries:

    http://viz-beta.u-strasbg.fr/
    http://cdsarc.u-strasbg.fr/

these are locations at which at least sometimes a non-production
Vizier instance is running.  Not really intended for end users,
but may be useful for diagnostics, debugging etc.  Don't advertise
it in the change log.