ttools: parallelise most Plotter implementations
Rework the various Plotter implementations so that their direct
or indirect access to TupleSequences is done within a SplitCollector,
and executed using the TupleRunner from the supplied DataStore.
This means that most plotting can be done in parallel if the
TupleRunner's concurrency policy determines it's a good idea.

The performance increases seem to be good, typically 6-7 times
speedup on my 8-core machine over the single-threaded implementation
for large datasets.  In some cases there will be more memory usage,
though in *most* cases not too much (i.e. not scaling with row count).

The implementation changes are mostly not too complicated.
This commit covers most of the Plotters but there are some which are not
readily parallelised for one reason or another and not tackled here:
SpectrogramPlotter, LinePlotter and Line3DPlotter.