topcat: fix GUI progress reporting for concurrency
Modify Progresser so that it can accept a smaller number of large
updates as alternative to a large number of small updates.
The GuiDataStore can then arrange to make a small number of updates,
so that they don't interfere with each other.  Without this change
there was much contention for the AtomicLong used as the Progresser's
state when it was used by many threads concurrently, which led to
very poor performance.

A previous attempt to mitigate this was substituting a LongAdder
for the AtomicLong, but it didn't help that much.  This solution
is better.