ttools: replace Iterable<double[]> with CoordSequence
A couple of methods like

   Iterable<double[]> createDataPosIterable(...)

have been replaced with

   Supplier<CoordSequence> createDataPosSupplier(...)

CoordSequence is a Splittable, so this makes it possible to
parallelise processing of those sequences.
Behaviour should be unaffected by this change.  The interface
change is to a fairly obscure part of the API so is not very
likely to cause inconvenience.

Also took the opportunity to reduce some duplication of code:
the DataPosIterator implementation code that was in both
ttools.plot2.SubCloud and topcat.plot2.GuiPointCloud has been
combined into new class ttools.plot2.DataPosSequence.