topcat: add healpix figure inclusion placeholder method
When drawing figures on the plot to indicate point inclusion,
it works for positions indicated by lon/lat coordinates,
but not for healpix indices.  I couldn't work out why this was,
but eventually tracked it down to SkyFigureMode and figured out
how to extend that so that healpix works as well.

I haven't actually done that, since it would require more methods
in various other classes, and I'm not sure it's worth the effort,
but I've added placeholder methods to make it more obvious how to
tackle this if I get round to it in the future.

No behaviour change at this commit.