ttools: parallelise PlotUtil.getClosestRow
This method is used for determining the closest row to a clicked
point in topcat activation.

Note that the PlotUtil.getClosestRow method now takes a
Supplier<TupleSequence> rather than a TupleSequence.
This makes a bit more sense and works better with the split
processing utility classes, though if there was any difficulty
with providing a (reusable) supplier, it could be changed back.

I have carried forward a thread interruption check from the
previous implementation, but I have to admit I don't know whether
it will operate when the collector is being run in the ForkJoinPool.