ttools: parallelise ranging in SpectrogramPlotter
Parallelise one part of the spectrogram processing, but not others.

For various reasons, the spectrogram is not a good candidate
for parallelisation given the current framework.
A spectrogram is not expected to have a huge number of rows,
the processing load comes from having many samples per row,
so the TupleProcessor rules of thumb about when to split
will probably not work well.  Also, the painting is done direct
to a graphics context, which cannot be done concurrently.
Both of these could be overcome with more effort if concurrent
spectrogram plotting looks important to have.