ttools: partially parallelise line plotters
LinePlotter and Line3dPlotter are problematic for parallelisation.
Where the sort sequence is supposed to be the input row sequence,
the SplitCollector machinery is not good, because it can do blocks
in any order.  In practice, a straightforward parallelisation of
natural-order line plotting would leave a gap between each block.
It would be possible to add hooks to the Splittable classes to be
able to reconstruct the order in which blocks were done and draw
in the connecting lines at the end, though quite fiddly.
Also, the lines are drawn directly to the graphics context,
which cannot be done concurrently.

However, some parts of the code are parallelised: aux ranging and
point sorting.