topcat: log uncaught throwables in PlotPanel
Some Runnables submitted to ExecutorServices in PlotPanel were
capable of throwing uncaught exceptions (Errors or RuntimeExceptions)
that just got swallowed by the execution system since the get()
method of the corresponding Future was never called.
This resulted (in test code) in some AssertionErrors happening
but never being seen.

Be more careful about how these jobs are submitted, so that the
exceptions are always at least reported through the logging system.