topcat: ensure blob completion callbacks happen
If the mask creation fails (which it shouldn't, but was happening
during development because of an uncaught AssertionError)
the blob painting status never gets reset, meaning the blob just
stays painted and can't be removed.  So be more careful to ensure
that the status reset action gets called even in case of failures.