ttools: add DataStore.getTupleRunner method
Store a TupleRunner inside a DataStore.  Although TupleRunner is
not part of the memory-sensitive storage that DataStore is designed
to take care of, a TupleRunner instance will usually be required
in the same places as a DataStore instance, so it makes sense to
keep them together.  This arrangement results in quite minor
changes to API methods.