ttools: add parallel variant of plot2-examples test
Add a new build file target plot2-examples-par alongside plot2-examples.
These targets build the figures for SUN/256 and SUN/253, but also
provide a useful regression test - figures should be pixel-for-pixel
equal between commits unless there's some comprehensible reason why.

The plot2-examples-par target forces (in most cases) parallel execution
of these figure generation plotting commands, so testing these outputs
against the default (generally sequential) ones is a useful check for
the parallelisation.  In some cases discrepancies are expected,
for instance for the Aux shading mode in which the output depends
on the (unimportant) order in which points are plotted.

Note that neither of these targets is currently invoked automatically
as part of the build or test process, since they are quite slow.