topcat: change coverage icon for HipsSelector
Use an elliptical rather than circular tiny icon to indicate
sky coverage for HiPS surveys - I think the Aitoff-like shape
suggests the sky better than the circle.  The coverage region
is not carefully calculated, so it gives a slightly misleading
quantitive idea of the coverage fraction (could be fixed if I
could be bothered to do the maths), but I think it's enough to
give an idea.

Also make some updates to the Hips2fits activation actions
in SUN/253.