topcat: update documentation screenshot montage
The cover image for the manual was extremely out of date, with only
pre-v4 plot windows.  I've completely redone it.

The mini version is the result of a GIMP blur operation followed by
a 1/4 shrink.  The fullsize version has a white rather than transparent
background - this is unfortunately necessary because otherwise the
PDF generation makes the background look black on a white page.
That's probably a bug in the extremely old Fop version I'm using,
and there may be some way round it, but I haven't worked out what.