topcat: update rowMap on subset expression edit
When the JEL expression defining the current Row Subset was edited
in the Subsets window, it wasn't updating the view table's rowMap
array; this meant that the Apparent row count didn't update in the
main control window and if the table data window was visible, the
row display there wasn't updated either.  Fix it so the updates
take place properly.