xdoc: update XSLT <javadoc> processing for Java 8
The toHTML1.xslt stylesheet turns xdoc <javadoc> elements into hyperlinks;
when the reference is to a method (member='func(type,type)') then
it generates an href attribute pointing to the relevant part of
the javadocs.  However, it seems that the form that javadoc
command uses for these anchors in the generated javadocs has
changed between java 6 and java 8.
Java 6 had anchors like "#func(type, type)', but
java 8 has "#func-type-type-".
Modify the hyperlink output of toHTML1.xslt to use the java 8 form.
This has been done ad hoc to fit the members I've come across so far;
it's possible that further changes would be required if some references
outside of current experience were required.