topcat: fix null DataStore plotting bug
When opening a plot window where topcat isn't able to guess what
plot to make (e.g. if there are too few numeric columns), the plot
was failing: empty plot panel (not even axes) and a NullPointerException
logged.  This was because it was attempting to use a DataStore
that had not been initialised.  Fix it by initialising the DataStore
with a dummy (non-null) value that behaves properly.

This bug was introduced during the multi-threading changes, so it
only appeared in the last public release.