vo: permit blank RA/Dec fields in SSA query window
None of the query parameters are mandatory in SSA (1.04 and 1.1 both),
so the user ought to be permitted to supply blank values for RA and Dec
as well as for Diameter (the latter is already permitted).
This is in fact reasonable user behaviour if the service contains
theoretical rather than observed spectra, a fact which I probably
missed when I originally wrote it with these fields required.
Fix it so that if RA and Dec are blank (though not just one of them),
the query is still submitted.

This change was in response to a request by Petr Skoda.