ttools: maximum marker size goes up from 5 to 9
The largest marker size provided in the plot2 GUI selector is
now 9 instead of 5.  A user requested larger values, and with
higher screen pixel densities it makes sense.
The only reason it can't easily get bigger than this is that
the representation in the JComboBox doesn't show up properly
for larger values.  If people want bigger numbers, do something
about that.