ttools: split PlotScene out of PlotDisplay
Refactor PlotDisplay so that it only contains the Swing-specific
parts of the display, delegating the generic parts of the job
to a new class PlotScene.  PlotScene can then be used as a basis
for non-Swing-based live plots.

PlotScene is in the ttools.plot2 rather than ttools.plot.task package,
since it is not specific to STILTS commands, so the related
PlotCaching class moves there too.  PlotScene now has no dependencies
on classes in plot2.ttools.task.

This commit does represent a change to the ttools.plot2 API.