topcat: use configureLogging from InvokeUtils
The InvokeUtils implementation of the configureLogging function
is almost identical to the one contained here, so use that instead
of copying the code.  I think the differences are down to earlier
misconceptions I had about the logging system, so I believe
behaviour will be unchnanged, but maybe keep an eye on it.