ttools: tweak actual return type of getCoords array
The getCoords method in CartesianVectorCoordSet was returning
a FloatingCoord[] array rather than the declared CoordArray.
That shouldn't matter, but it was causing an invocation of
the slightly dodgy PlotUtil.arrayConcat utility method to fail
(at build time) elsewhere under some changes.
This fix isn't that clean, buy it makes the problem go away for now.