ttools: implement new RGB Shading Mode (not used)
This shading mode plots shapes in a colour given by an RGB coordinate,
an integer interpreted as RRGGBB bytes.  Is this useful?  I don't know.
It doesn't attempt to merge multiple pixels overplotted to average
the RGB values, which it probably should do.

For now this is not offered as one of the options for 2D or 3D plots,
but the implementation is there, so it would be a one-liner to do that
(just add IRGB to MODES_2D/MODES_3D arrays).