tamfits: use long int for ASCII table I9+, not I10+
When processing the ASCII table header, integers that are declared
with TFORM In, were treated as int for n<=10 and as long for n>10.
This is legal according to the standard, but I10 values can exceed
the 32-bit limit, so the behaviour is changed to use long for n==10
as well as n>10.

This issue was reported by Mark Bentley (ESA).