table: modify xmatch to bin using HashSet not TreeSet
The row binning at the heart of the table.join crossmatching
puts lots of row links in a LinkSet, which eventually has to be sorted.
Very early on, I based the default implementation on a
java.util.TreeSet, so that the sorting came for free.
Some recent benchmarks however suggest this is a pretty bad idea,
since TreeSet has O(log N) time for most operations in comparison
with HashSet constant time.  For certain large matches (ones with
lots of at least potential links) a LinkSet based on a HashSet
performs significantly better.

There are opportunities here to mess up the output result sequence,
but I *think* I've avoided them.  Other than that, there shouldn't
be any change to results.