ttools: modify taplint QUERY_STATUS processing
Modify the way that taplint behaves if a VOTable result contains
a pre-table <INFO name="QUERY_STATUS" value="OVERFLOW"/>.
Previously OVERFLOW declaration was only permitted in the post-table
declaration, but now it's permitted in the pre-table one as well,
as permitted by DALI 1.1.
Tighten up some of the rest of the QUERY_STATUX logic too.

This issue was reported by Gregory Dubois-Felsmann (IPAC).