topcat: use Swing generics properly
Parameterise use of JComboBox and JList etc everywhere,
so that most SuppressWarnings annotations can be removed.
The exception is classes in the package uk.ac.starlink.topcat.plot,
which are mostly moribund (old-style plot windows, still available
but now deprecated and not being actively maintained).

Quite a lot of code is touched, and some of the changes are more
than just textual, so there is possibility for regression issues.
Think it's OK though.