ttools: modify Coord input-storage conversion API
Replace the Coord.inputToStorage method by a factory method that
returns a function.  The benefit is that the user->storage mapping
can be set up in a way that is dependent on the relevant
column metadata.  This makes e.g. FloatingArrayCoord implementation
a bit more efficient, and also gives more flexibility to set up
coordinate storage for more complex input data types such as polygons.