ttools: remove DomainMapper member from Coord & Input
The DomainMapper attribute of Coord instances was mostly unused,
only required for the TimeCoord.  But recent redesign of Coord.inputStorage
method means it's no longer used.  It's limited use suggests that its
introduction in the first place may not have been very good design.
This change simplifies things a bit.