table: add Tables.getXtype() method
The VOTable extended type attribute can now be retrieved with the
generic method Tables.getXtype().  The xtype has been around for
a while, but DALI 1.1 uses it more heavily than before, so it's
useful to have a standard way to retrieve this, rather than
using VOTable-specific code.  Possibly in future versions this
may be implemented in terms of a ValueInfo.getXtype method
(which is what happened to getUtype).