ttools: fix skygrid plot from stilts to see viewsys
The SkyGrid plotter was not being fed the plot's View SkySystem
from the STILTS context (though it was from TOPCAT).  The effect
was that the stilts plot2sky viewsys parameter was ignored
(effectively assumed equatorial), making the skygrid plot from stilts
incorrect for non-equatorial viewsys values.  Oops.
Make the necessary changes to the stilts plotting harness, and also
add some checks and comments in the SkyGridPlotter that would spot
if this kind of thing were to arise again.