ttools: Coord.inputStorage has new DomainMapper[] arg
Re-introduce the DomainMappers argument (in a somewhat modified form)
to the Coord method that determines how data values are stored.
Apart from the UI work required to acquire the DomainMapper choices
from user preferences, this is the last piece of the puzzle required
to allow different interpretations under user control of the input
column data to be stored as plottable data coordinate values.