topcat: slightly improve logging in PlotPanel
Exceptions thrown during plots are now written to the console;
however this isn't done properly through the logging system.
I think doing it properly would require modification to task.InvokeUtils.

I thought that used to work (write stacktraces if -debug was set),
but it doesn't seem to do it now; the corresponding functionality
in STILTS is handled explicitly by LineInvoker.
Possibly come back to this and implement a better fix later.