topcat: provide DomainMapper selection GUI
Selection of Coordinate values with non-trivial Domains now
provide also a DomainMapper selector.  That means that you can
both see what mapping is in use, and choose a mapping to order.
This makes the TimePlot finally usable for tables in which the
metadata does not explicitly mark up time systems (or in which
that marking up is not recognised).