fits: read String values with readBytes
Use new bulk BasicInput method readBytes rather than multiple readByte
calls when reading string values.  This can be quite a bit faster for
long strings (thousands of characters).  It looks like it might be
slightly slower for short strings, but that effect seems to be small.

A similar change could be made for reading numeric arrays,
it may be worthwhile where large arrays are being read.
That would touch quite a bit more code in ColumnReader,
I haven't attempted it so far.