ttools: add PolygonMode STAR option
This draws lines from the nominal center of the shape to each vertex.
Sometimes it's a useful way to see regions.  There is a cost however
for all modes, since it means that the nominal center has to be
acquired and passed around in any case.  In most cases I'd expect
that to be much cheaper than the actual plotting though, so it's
probably OK.