ecsv: add snakeyaml library
SnakeYaml is a YAML 1.1 I/O library that seems to be under
active maintenance.  At 300kbyte, it is no doubt overkill
for what we need here, namely a YAML parser, but at least
it doesn't have any external dependencies.  It would be possible
to strip it down to only a few of the classes, but YAML seems
pretty complicated, so it would need to be done with care.

I tried JYaml, which is much smaller (60kbyte), but maintenance seems
to be moribund, and it didn't work for the basic Astropy example.ecsv
file.  JVyaml (150kbyte) is another possibility, but again it's long
out of maintenance.

This commit adds version 1.25 of SnakeYaml, acquired from
https://repo1.maven.org/maven2/org/yaml/snakeyaml/1.25/snakeyaml-1.25.jar.