ttools: use HealpixNestedBMOCs in MOC manipulation
When replacing the PixTools implementation of HealpixImpl with
the CDS implemetation, there are more opportunities to do things
cleverly.  Modify the implementation of a couple of MOC manipulation
algorithms to make use of the fact that the CDS implementation
exposes BMOCs rather than just a list of the lowest-order tiles.
This should make things more efficient.