table: remove PixTools library from runtime
The PixTools library of Nickolai Kuropatkine is no longer used anywhere
in TABLE/STILTS/TOPCAT; it has been replaced with the cds-healpix-java
library of F-X Pineau.  The jar files are removed from the build,
though retained for comparison in the classpath used for unit
testing of the TTOOLS package.

Note the move from PixTools to cds-healpix-java in the topcat
and stilts user document change logs.