ttools: fix sortaxis option for Line plot in Time
The LinePlotter sortaxisN options when the LinePlotter is used
in the TimePlotType are now TIME or None, rather than X, Y or None.

I thought about defaulting it to TIME in the time plot, since one
will normally want these points to be in sequence, but the
performance hit is non-negligable, and in many cases (if the
table is in order anyway) it won't make any visible difference,
so leave this to the user after all.