topcat: STILTS window now reports domain mappers
The <coordname>typeN parameters that relate to chosen DomainMappers
are now correctly reported in the command displayed by topcat's
STILTS window.  Previously the chosen DomainMapper was just ignored
for this purpose, which meant that selecting e.g. DecYear rather
than MJD for a time plot in the GUI was not reflected in the reported
STILTS command, so the stilts plot wouldn't show the right thing.

This change should have been made when the DomainMapper changes were
made a few commits back, but I didn't think about it.
But there has not been a public release in the mean time.