ttools: fix Range.limit(lo,hi) method
The limit(lo,hi) method of plot.Range was not setting the
upper limit if the lower limit was blank.  I'm fairly sure that
was an oversight, so I've fixed it to make it work as I'd expect.

I'm a bit nervous about this since Range has been in place for a
very long time and it's in use in many places.  But hopefully
there will be no problems.