ttools: make time plot default ranges more sensible
When doing a time plot with no range either explicitly supplied
or inferred from data, the range on the time axis was 0-1,
i.e. 1970-01-01T00:00:00 - 1970-01-01T00:00:01, which wasn't
very human friendly.  Fix it so that it uses more sensible ranges:
the current epoch is used as one or other limit, and the extent
is (at least typically) of the order of years.