topcat: adjust default marker size by table row count
The initial default size of scatter plot markers (StyleKeys.SIZE)
when plotting a table is now sensitive to the number of rows in
that table.  This is a fairly blunt instrument, but it means
that if you're plotting a table with a few tens of rows the plot
will initially have big chunky markers, but if it has a million
rows the markers will be small.  The size currently doesn't go
any smaller than the previous default (size=1).

This recovers, more or less, some behaviour that was in TOPCAT v3
visualisation.  Note this change has not been replicated in STILTS,
since it's harder to hack the defaulting in stilts, and also since
in stilts the row count may not be known.  Those considerations
(wanting the behaviour to be identical in topcat and stilts) are
probably why I didn't implement this nice usability feature when
I initially wrote the v4 visualisation.