ttools: recognise STC-S areas by xtype
Recognise xtype="stc-s" or xtype="stc" (case-insensitive) as
indicating STC-S content.  I don't think that this is standardised
anywhere for now, but it can't hurt, since these xtypes could
be used ad-hoc if required, and could hardly be intended to mean
anything else.

Markus says that xtype="ADQL:region" is common for STC-S,
and that it might possibly be standards-based (where?),
so I did consider adding that as well.
However, columns that are marked like that will probably be
s_regions that I'll catch by name or utype in any case,
and "ADQL:region" could conceivably be used to label a MOC
as well, so don't do that for now.