table: replace Math.asin with FastMath.asin in SkyMatchEngine
Evaluations of arcsin are in some cases (following the switch to
cds-healpix-java) the dominant computational cost in sky crossmatching.
Following advice from F-X Pineau, replace invocations of Math.asin()
with FastMath.asin() {Abstract,Ellipse}SkyMatchEngine.
This implementation is found in the cds-healpix-java package,
but comes originally from Apache FastMath.
It's much faster than the java.lang.Math version, and according to
its own documentation it's more accurate as well.  The differences
between the two implementations certainly seem to be very small
(as a newly-added simple unit test confirms), and other unit and
regression tests don't show any detectable difference in crossmatching
behaviour.  I'm slightly nervous about this change to the numerics,
but if it's good enough for F-X, it's probably good enough for me.

At this stage in the interests of transparency I have not switched
the implementation for asin calls elsewhere in the code;
I will consider that if it looks like they become significant
computational costs.